Refactor logic, clean up code
